Change Events to be component and add on HomePage
