+++ apps/web-app/pages/index.tsx
@@ -1,36 +1,24 @@
+// pages/index.tsx
 import { GetServerSideProps } from "next"
-import axios from "axios"
-
-import { EventsDTO, SessionsDTO } from "../types"
-import CalendarPage from "../templates/CalendarPage"
+import { EventsDTO } from "../types"
+import HomeTemplate from "../templates/Home"
 
 type Props = {
     events: EventsDTO[]
-    sessions: SessionsDTO[]
 }
 
-export default function Event({ sessions, events }: Props) {
-    return <CalendarPage sessions={sessions} events={events} />
-}
+const Home = ({ events }: Props) => <HomeTemplate events={events} />
 
-export const getServerSideProps: GetServerSideProps = async ({ req, res }) => {
+export default Home
+
+export const getServerSideProps: GetServerSideProps = async ({ res }) => {
     try {
         const url = process.env.URL_TO_FETCH
 
         const eventsResponse = await fetch(`${url}/api/fetchEvents`)
-
-        const events: EventsDTO[] = await eventsResponse.json()
-
-        const responseSessions = await axios.get(`${url}/api/fetchSessions`, {
-            headers: {
-                Cookie: req.headers.cookie || "" // Pass cookies from the incoming request
-            }
-        })
-
-        const sessions = await responseSessions.data
-
+        const events = await eventsResponse.json()
         return {
-            props: { sessions, events }
+            props: { events }
         }
     } catch (error) {
         res.statusCode = 404
